data/database/requirements.txt
pandas>=1.3.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.2